+++ apps/tabler-app/src/queries/GetClubsMapQuery.ts
@@ -1,17 +1,33 @@
 import gql from 'graphql-tag';
+import { AddressFragment } from './AddressFragment';
 
 export const GetClubsMapQuery = gql`
   query ClubsMap {
-    Clubs {
-      id
-      name
-      logo
-      club
+        Clubs {
+            id
+            name
+            logo
+            club
 
-      location {
-        longitude
-        latitude
-      }
-    }
+            meetingplace1 {
+                ...AddressFragment
+            }
+
+            meetingplace2 {
+                ...AddressFragment
+            }
+
+            info {
+                first_meeting
+                second_meeting
+            }
+
+            location {
+                longitude
+                latitude
+            }
+        }
   }
+
+  ${AddressFragment}
 `;